Add soln folder, start test file
